Add the units and description attributes to field attlists
This commit adds parsing in Registry to extract the units and
description attributes from a variable definition in Registry and add
them as attributes to a field definition.

This commit also modifies the attribute lists to allow each constituent
from a var_array to have it's own unique attributes.

Finally, this commit updates each core's core_interface to allow
attributes to be added automatically by registry generated code.